requirements.txt
annotated-types==0.7.0
anthropic==0.53.0
anyio==4.9.0
appdirs==1.4.4
asyncio==3.4.3
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
notion-client==1.0.0
pdoc3==0.11.6
playwright==1.52.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dotenv==1.0.0
requests==2.32.3
shiboken6==6.9.1
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
types-psycopg2==2.9.21.20250318
types-setuptools==80.3.0.20250505
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==1.26.20
websockets==10.4
zipp==3.22.0
